chore: split integration tests from component test file
